Switched to 'yauzl' from 'unzip' to minimize deps
The 'unzip' NPM package was reliant on many other packages, some of
which were deprecated; this makes it less reliable than 'yauzl'. which
only has two low-level dependencies.

Minimizing deps will simulatneously improve package size, security, and
resolve issue #3.
