+++ src/install-terraform/set-perms.js
